+++ src/bellman_ford/CMakeLists.txt
@@ -3,4 +3,6 @@ ADD_LIBRARY(bellman_ford OBJECT
     bellman_ford_driver.cpp
     #bellman_ford_neg.c
     #bellman_ford_neg_driver.cpp
+    edwardMoore.c
+    edwardMoore_driver.cpp
     )
