--- src/edwardMoore/CMakeLists.txt
@@ -1,4 +1,0 @@
-ADD_LIBRARY(edwardMoore OBJECT
-    edwardMoore.c
-    edwardMoore_driver.cpp
-    )
